Add merge utiliy functions and add input path completion for score
